+++ src/main/java/com/banchan/dto/Votes.java
@@ -1,5 +1,6 @@
 package com.banchan.dto;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 
 import javax.persistence.*;
@@ -15,6 +16,7 @@ public class Votes {
     @Column(name = "userId", nullable = false)
     private int userId;
 
+    @JsonIgnore
     @ManyToOne
     @JoinColumn(name = "questionId", nullable = false)
     private Questions question;
